some problem fixed

+++ grails-app/controllers/org/auction/AccountController.groovy
@@ -11,22 +11,21 @@ import grails.plugin.springsecurity.annotation.Secured
 class AccountController {
 
     def springSecurityService= new SpringSecurityService()
+
+    @Secured(["IS_AUTHENTICATED_FULLY"])
+    def index(){
+        User user=springSecurityService.currentUser
+        Account account= Account.findByUsername(user.username)
+        params.id = account.id
+    }
+
     @Transactional
 	def create() {
 		respond new Account(params)
 	}
 
-    def show() {
-        if(springSecurityService?.isLoggedIn()){
-            User user= springSecurityService.currentUser;
-            Account account= Account.findByUsername(user.username);
-            if(!account){
-                redirect controller: "login", action:"auth"
-            }
-            respond account
-        }else{
-            redirect controller: "login", action:"auth"
-        }
+    def show(Account accountInstance) {
+        respond accountInstance
     }
 
 	@Transactional
@@ -39,11 +38,10 @@ class AccountController {
             respond accountInstance.errors, view:'create'
             return
         }
-
         User user = new User(username:accountInstance.username, password: accountInstance.password)
         user.id=accountInstance.id
         if (!accountInstance.validate() || !user.validate()) {
-            flash.error = "The username or email is already exist!"
+            flash.message = "The username or email is already exist!"
             redirect action: "create", controller:"account"
         }else{
             accountInstance.save flush:true
@@ -59,16 +57,16 @@ class AccountController {
 	}
 
     @Secured(["IS_AUTHENTICATED_FULLY"])
-	def edit() {
-        if(springSecurityService.isLoggedIn()){
-            User user= springSecurityService.currentUser;
-            Account account= Account.findByUsername(user.username);
-            if(!account){
-                redirect controller: "login", action:"auth"
-            }
+	def edit(Account accountInstance) {
+        User user= springSecurityService.currentUser
+        Account account= Account.findByUsername(user.username)
+        def id=account.id
+        println id+"id"+accountInstance.id
+        if(!id.equals(accountInstance.id)){
+            redirect controller: "login", action:"denied"
+        }
+        else{
             respond account
-        }else{
-            redirect controller: "login", action:"auth"
         }
 	}
 
@@ -103,20 +101,26 @@ class AccountController {
 	}
 
     @Secured(["IS_AUTHENTICATED_FULLY"])
-	def delete() {
-        User user= springSecurityService.currentUser;
-        Account accountInstance= Account.findByUsername(user.username);
+	def delete(Account accountInstance) {
 		if (accountInstance == null) {
 			notFound()
 			return
 		}
+        User user= springSecurityService.currentUser;
+        Account account = Account.findByUsername(user.username);
+        def id = account.id
 
-        Collection<UserRole> userRoles = UserRole.findAllByUser(user);
-        userRoles*.delete()
-        user.delete()
-		accountInstance.delete flush:true
-
-		redirect controller:"logout", method: "POST"
+        if(!id.equals(accountInstance.id)){
+            redirect controller: "login", action:"denied"
+        }
+        else
+        {
+            Collection<UserRole> userRoles = UserRole.findAllByUser(user);
+            userRoles*.delete()
+            user.delete()
+    		accountInstance.delete flush:true
+		    redirect controller:"logout", method: "POST"
+        }
 	}
 	protected void notFound() {
 		request.withFormat {
